Concise commit message:
```text
feat: add watermark method to NexusMemoryCore

- Add `watermark` function to record LDO action IDs and timestamps.
```

+++ app/src/main/java/dev/aurakai/auraframefx/domains/aura/chronokineticforge/engines/RebelliousPaintDripEngine.kt
@@ -464,7 +464,7 @@ object RebelliousPaintDripEngine {
         val x = stream.startX * canvasSize.width + cos(stream.angle) * distance * canvasSize.width
         val y = stream.startY * canvasSize.height + sin(stream.angle) * distance * canvasSize.height
 
-        return Offset(x.toFloat(), y.toFloat())
+        return Offset(x, y)
     }
 
     // ═════════════════════════════════════════════════════════════════
